+++ library/common/http/dispatcher.h
@@ -11,7 +11,7 @@
 
 #include "common/common/logger.h"
 
-#include "library/common/include/c_types.h"
+#include "library/common/types/c_types.h"
 
 namespace Envoy {
 namespace Http {
