+++ library/common/http/header_utility.h
@@ -3,7 +3,7 @@
 #include "envoy/buffer/buffer.h"
 #include "envoy/http/header_map.h"
 
-#include "library/common/include/c_types.h"
+#include "library/common/types/c_types.h"
 
 namespace Envoy {
 namespace Http {
