+++ library/common/types/c_types.c
@@ -1,4 +1,4 @@
-#include "library/common/include/c_types.h"
+#include "library/common/types/c_types.h"
 
 #include <string.h>
