+++ test/common/http/dispatcher_test.cc
@@ -11,7 +11,7 @@
 #include "gtest/gtest.h"
 #include "library/common/http/dispatcher.h"
 #include "library/common/http/header_utility.h"
-#include "library/common/include/c_types.h"
+#include "library/common/types/c_types.h"
 
 using testing::_;
 using testing::NiceMock;
