+++ test/common/http/header_utility_test.cc
@@ -2,7 +2,7 @@
 
 #include "gtest/gtest.h"
 #include "library/common/http/header_utility.h"
-#include "library/common/include/c_types.h"
+#include "library/common/types/c_types.h"
 
 namespace Envoy {
 namespace Http {
